+++ src/lib/pages/tx-details/components/tx-message/TxMsgExpand.tsx
@@ -41,7 +41,7 @@ export const TxMsgExpand = ({
 
   const { "@type": type, ...body } = msgBody;
   const isIBC =
-    Boolean(log.events.find((event) => event.type === "send_packet")) ||
+    Boolean(log?.events.find((event) => event.type === "send_packet")) ||
     type.startsWith("/ibc");
   const isOpinit = Boolean(type.startsWith("/opinit"));