+++ src/lib/utils/tx/extractTxLogs.ts
@@ -3,26 +3,23 @@ import type { Event, logs } from "@cosmjs/stargate";
 import type { TxResponse } from "lib/services/tx";
 
 export const extractTxLogs = (txData: TxResponse): logs.Log[] => {
-  const msgLogs = txData.tx.body.messages.map((_, idx) => ({
-    msg_index: idx,
-    log: "",
-    events: [] as Event[],
-  }));
+  if (txData.logs.length > 0) return txData.logs;
 
-  if (txData.logs.length > 0)
-    txData.logs.forEach((log) =>
-      msgLogs[log.msg_index].events.push(...log.events)
-    );
-  else
-    txData.events.forEach((event) => {
-      const attribute = event.attributes.find(
-        (attr) => attr.key === "msg_index"
-      );
-      if (attribute) {
-        const index = Number(attribute.value);
-        msgLogs[index].events.push(event);
+  const logs: Record<string, logs.Log & { events: Event[] }> = {};
+  txData.events.forEach((event) => {
+    const attribute = event.attributes.find((attr) => attr.key === "msg_index");
+    if (attribute) {
+      const index = attribute.value;
+      if (!Object.keys(logs).includes(attribute.value)) {
+        logs[index] = {
+          msg_index: Number(index),
+          log: "",
+          events: [] as Event[],
+        };
       }
-    });
+      logs[index].events.push(event);
+    }
+  });
 
-  return msgLogs;
+  return Object.values(logs).toSorted((a, b) => a.msg_index - b.msg_index);
 };