+++ CHANGELOG.md
@@ -41,6 +41,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Improvements
 
+- [#750](https://github.com/alleslabs/celatone-frontend/pull/750) api v1 - recent codes list
+
 ### Bug fixes
 
 ## v1.5.1