+++ CHANGELOG.md
@@ -44,6 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Improvements
 
 - [#750](https://github.com/alleslabs/celatone-frontend/pull/750) api v1 - recent codes list
+- [#752](https://github.com/alleslabs/celatone-frontend/pull/752) Support contract state's key as base64
 
 ### Bug fixes